Currency conversion:  - Currency enum added;  - ExchangeRate class added to represent exchange rate between 2 currencies;  - CurrencyDao interface and HttpCurrencyDao implementation (to serve the purpose of getting rates from NBU resource) are added;  - CurrencyService interface and CachedCurrencyService implementation are added;  - Converters class added: it contains different implementation of Converter interface: CurrencyConverter and RoundingConverter that can be gathered to chain of responsibility;  - GetMovieRequest extended to keep currency that comes from http get request;  -  MovieService.getMovieById() changed in part of parameters list: ite receives GetMovieRequest now;

+++ src/main/java/edu/ted/web/movieland/dao/MovieDao.java
@@ -11,5 +11,5 @@ public interface MovieDao {
     List<Movie> findAll(Sorting sorting);
     List<Movie> getNRandomMovies(int count);
     List<Movie> getMoviesByGenre(int genreId, Sorting sorting);
-    Optional<Movie> getMovieById(int movieId);
+    Optional<Movie> getMovieById(long movieId);
 }
